+++ legadilo/users/urls.py
@@ -25,6 +25,6 @@ urlpatterns = [
     path("~settings/", view=views.user_update_settings_view, name="update_settings"),
     path("<int:pk>/", view=views.user_detail_view, name="detail"),
     path("notifications/", views.list_notifications_view, name="list_notifications"),
-    path("tokens/", views.list_tokens_view, name="list_tokens"),
+    path("tokens/", views.manage_tokens_view, name="manage_tokens"),
     path("tokens/<int:token_id>/delete/", views.delete_token_view, name="delete_token"),
 ]