+++ legadilo/users/views/__init__.py
@@ -14,7 +14,7 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from .manage_tokens_views import delete_token_view, list_tokens_view
+from .manage_tokens_views import delete_token_view, manage_tokens_view
 from .notifications_views import list_notifications_view
 from .user_views import (
     user_detail_view,
@@ -26,7 +26,7 @@ from .user_views import (
 __all__ = [
     "delete_token_view",
     "list_notifications_view",
-    "list_tokens_view",
+    "manage_tokens_view",
     "user_detail_view",
     "user_redirect_view",
     "user_update_settings_view",