+++ legadilo/utils/testing.py
@@ -36,3 +36,7 @@ def serialize_for_snapshot(value: Any) -> str:
 
 def all_model_fields_except(model: type[models.Model], excluded_fields: set[str]):
     return [field.name for field in model._meta.fields if field.name not in excluded_fields]
+
+
+def build_bearer_header(jwt: str = ""):
+    return f"Bearer {jwt}"